Rename to `capybara_accessibility_audit`
Replace all occurrences of:

* `CapybaraA11y` in Ruby code with `CapybaraAccessibilityAudit`
* `capybara_a11y` in Ruby code with `capybara_accessibility_audit`
* `capybara_a11y` in file and directory names with `capybara_accessibility_audit`

Once merged, we'll need to rename the repository as well.
