--- lib/capybara_a11y/version.rb
@@ -1,3 +1,0 @@
-module CapybaraA11y
-  VERSION = "0.1.0"
-end
