+++ lib/capybara_accessibility_audit.rb
@@ -2,7 +2,7 @@ require "zeitwerk"
 loader = Zeitwerk::Loader.for_gem
 loader.setup
 
-module CapybaraA11y
+module CapybaraAccessibilityAudit
   # Your code goes here...
 end
