+++ lib/capybara_accessibility_audit/version.rb
@@ -1,0 +1,3 @@
+module CapybaraAccessibilityAudit
+  VERSION = "0.1.0"
+end
