--- test/capybara_a11y_test.rb
@@ -1,7 +1,0 @@
-require "test_helper"
-
-class CapybaraA11yTest < ActiveSupport::TestCase
-  test "it has a version number" do
-    assert CapybaraA11y::VERSION
-  end
-end
