+++ test/capybara_accessibility_audit_test.rb
@@ -1,0 +1,7 @@
+require "test_helper"
+
+class CapybaraAccessibilityAuditTest < ActiveSupport::TestCase
+  test "it has a version number" do
+    assert CapybaraAccessibilityAudit::VERSION
+  end
+end
