+++ test/dummy/config/application.rb
@@ -17,7 +17,7 @@ require "rails/test_unit/railtie"
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
 Bundler.require(*Rails.groups)
-require "capybara_a11y"
+require "capybara_accessibility_audit"
 
 module Dummy
   class Application < Rails::Application
